+++ tests_cairo/math/mod.rs
@@ -1,2 +1,3 @@
+mod math_cmp_test;
 mod math_test;
 mod math_test_utils;